Combine src/background/{index.ts, install.ts}
The code in `install.ts` is run not just when the extension is installed but
whenever the service worker is initialized. Make the structure clearer by
combining the code from the Service Worker entry point and install.ts into one
module.
